Signup page adapted to the new Bootstrap theme \nHi from the template commit-msg hook
